Move variation function to Elliptic
- as suggested by Markus
- this makes the symv function slower by 1 memop per element
