+++ inc/dg/elliptic.h
@@ -19,44 +19,47 @@ namespace dg
 {
 
 /**
- * @brief A 2d negative elliptic differential operator \f$ -\nabla \cdot ( \chi \nabla ) \f$
+ * @brief A 2d negative elliptic differential operator \f$ -\nabla \cdot ( \mathbf{\chi}\cdot \nabla ) \f$
  *
  * @ingroup matrixoperators
  *
- * The term discretized is \f[ -\nabla \cdot ( \chi \nabla ) \f]
- * where \f$ \nabla \f$ is the two-dimensional nabla and \f$\chi\f$ is a
- * (possibly spatially dependent) tensor.
- * In general coordinates that means
+ * The term discretized is \f[ -\nabla \cdot ( \mathbf{\chi} \cdot \nabla ) \f] where \f$
+ * \nabla \f$ is the two-dimensional nabla and \f$\chi = \sigma
+ * \mathbf{\tau}\f$ is a (possibly spatially dependent) tensor with scalar part
+ * \f$ \sigma\f$ (usually the volume form) and tensor part \f$ \tau\f$ (usually
+ * the inverse metric). In general coordinates that means
  * \f[ -\frac{1}{\sqrt{g}}\left(
  * \partial_x\left(\sqrt{g}\left(\chi^{xx}\partial_x + \chi^{xy}\partial_y \right)\right)
  + \partial_y\left(\sqrt{g} \left(\chi^{yx}\partial_x + \chi^{yy}\partial_y \right)\right) \right)\f]
  is discretized.
- Per default, \f$ \chi\f$ is the metric tensor but you can set it to any tensor
+ Per default, \f$ \chi = \sqrt{g} g^{-1}\f$ but you can set it to any tensor
  you like (in order for the operator to be invertible \f$\chi\f$ should be
  symmetric and positive definite though).
+
  Note that the local discontinuous Galerkin discretization adds so-called jump terms
  \f[ D^\dagger \chi D + \alpha \chi_{on/off} J \f]
  where \f$\alpha\f$  is a scale factor ( = jfactor), \f$ D \f$ contains the discretizations of the above derivatives, and \f$ J\f$ is a self-adjoint matrix.
  (The symmetric part of \f$J\f$ is added @b before the volume element is divided). The adjoint of a matrix is defined with respect to the volume element including dG weights.
  Usually, the default \f$ \alpha=1 \f$ is a good choice.
  However, in some cases, e.g. when \f$ \chi \f$ exhibits very large variations
- \f$ \alpha=0.1\f$ or \f$ \alpha=0.01\f$ might be better values. 
+ \f$ \alpha=0.1\f$ or \f$ \alpha=0.01\f$ might be better values.
  In a time dependent problem the value of \f$\alpha\f$ determines the
- numerical diffusion, i.e. for too low values numerical oscillations may appear. 
+ numerical diffusion, i.e. for too low values numerical oscillations may appear.
  The \f$ \chi_{on/off} \f$ in the jump term serves to weight the jump term with \f$ \chi \f$. This can be switched either on or off with off being the default.
  Also note that a forward discretization has more diffusion than a centered discretization.
 
  The following code snippet demonstrates the use of \c Elliptic in an inversion problem
  * @snippet elliptic2d_b.cu invert
  * @copydoc hide_geometry_matrix_container
- * This class has the \c SelfMadeMatrixTag so it can be used in blas2::symv functions
+ * This class has the \c SelfMadeMatrixTag so it can be used in \c blas2::symv functions
  * and thus in a conjugate gradient solver.
- * @note The constructors initialize \f$ \chi=g\f$ so that a negative laplacian operator
- * results
- * @note The inverse of \f$ \chi\f$ makes a good general purpose preconditioner
+ * @note The constructors initialize \f$ \chi=\sqrt{g}g^{-1}\f$ so that a
+ * negative laplacian operator results
+ * @note The inverse of \f$ \sigma\f$ makes a good general purpose preconditioner
  * @note the jump term \f$ \alpha J\f$  adds artificial numerical diffusion as discussed above
+ * @note Since the pattern arises quite often (because of the ExB velocity \f$ u_E^2\f$ in the ion gyro-centre potential)
+ * this class also can compute the variation integrand \f$ \lambda^2\nabla \phi\cdot \chi\cdot\nabla\phi\f$
  * @attention Pay attention to the negative sign which is necessary to make the matrix @b positive @b definite
- *
  */
 template <class Geometry, class Matrix, class Container>
 class Elliptic
@@ -117,10 +120,8 @@ class Elliptic
         dg::assign( dg::create::inv_weights(g),   m_precond);
         m_temp = m_tempx = m_tempy = m_inv_weights;
         m_chi=g.metric();
-        m_vol=dg::tensor::volume(m_chi);
-        dg::tensor::scal( m_chi, m_vol);
+        m_sigma = m_vol = dg::tensor::volume(m_chi);
         dg::assign( dg::create::weights(g), m_weights_wo_vol);
-        dg::assign( dg::evaluate(dg::one, g), m_sigma);
     }
 
     /**
@@ -136,26 +137,47 @@ class Elliptic
         *this = Elliptic( std::forward<Params>( ps)...);
     }
 
-    ///@copydoc Elliptic3d::set_chi(const ContainerType0&)
+    /**
+     * @brief Change scalar part in Chi tensor
+     *
+     * Internally, we split the tensor \f$\chi = \sigma\mathbf{\tau}\f$ into
+     * a scalar part \f$ \sigma\f$ and a tensor part \f$ \tau\f$ and you can
+     * set each part seperately. This functions sets the scalar part.
+     *
+     * @param sigma The new scalar part in \f$\chi\f$
+     * @note The class will take care of the volume element in the divergence so do not multiply it to \c sigma yourself
+     * @attention If some or all elements of sigma are zero the preconditioner
+     * is invalidated and the operator can no longer be inverted. The symv
+     * function can still be called however.
+     * @tparam ContainerType0 must be usable with \c Container in \ref dispatch
+     */
     template<class ContainerType0>
     void set_chi( const ContainerType0& sigma)
     {
-        dg::blas1::pointwiseDivide( sigma, m_sigma, m_tempx);
+        dg::blas1::pointwiseDot( sigma, m_vol, m_sigma);
         //update preconditioner
-        dg::blas1::pointwiseDivide( m_precond, m_tempx, m_precond);
-        dg::tensor::scal( m_chi, m_tempx);
-        dg::blas1::copy( sigma, m_sigma);
+        dg::blas1::pointwiseDivide( m_inv_weights, sigma, m_precond);
+        // sigma is possibly zero, which will invalidate the preconditioner
+        // it is important to call this blas1 function because it can
+        // overwrite NaN in m_precond in the next update
     }
     /**
-     * @copydoc Elliptic3d::set_chi(const SparseTensor<ContainerType0>&)
-     * @note the 3d parts in \c tau will be ignored
+     * @brief Change tensor part in Chi tensor
+     *
+     * We split the tensor \f$\chi = \sigma\mathbf{\tau}\f$ into
+     * a scalar part \f$ \sigma\f$ and a tensor part \f$ \tau\f$ and you can
+     * set each part seperately. This functions sets the tensor part.
+     *
+     * @note The class will take care of the volume element in the divergence so do not multiply it to \c tau yourself
+     *
+     * @param tau The new tensor part in \f$\chi\f$ (must be positive definite)
+     * @note the 3d parts in \c tau will be ignored for 2d computations
+     * @tparam ContainerType0 must be usable in \c dg::assign to \c Container
      */
     template<class ContainerType0>
     void set_chi( const SparseTensor<ContainerType0>& tau)
     {
         m_chi = SparseTensor<Container>(tau);
-        dg::tensor::scal( m_chi, m_sigma);
-        dg::tensor::scal( m_chi, m_vol);
     }
 
     /**
@@ -179,9 +201,9 @@ class Elliptic
     /**
      * @brief Return the default preconditioner to use in conjugate gradient
      *
-     * Currently returns the inverse weights without volume elment divided by the scalar part of \f$ \chi\f$.
-     * This is especially good when \f$ \chi\f$ exhibits large amplitudes or variations
-     * @return the inverse of \f$\chi\f$.
+     * Currently returns the inverse weights without volume elment divided by the scalar part of \f$ \sigma\f$.
+     * This is especially good when \f$ \sigma\f$ exhibits large amplitudes or variations
+     * @return the inverse of \f$\sigma\f$.
      */
     const Container& precond()const {
         return m_precond;
@@ -244,33 +266,12 @@ class Elliptic
     template<class ContainerType0, class ContainerType1>
     void symv( value_type alpha, const ContainerType0& x, value_type beta, ContainerType1& y)
     {
-        multiply_sigma( alpha, 1., x, beta, y);
-    }
-
-    /**
-     * @brief Compute elliptic term with a possibly zero prefactor and add to output
-     *
-     * i.e this function computes \f[ y = -\alpha\nabla \cdot ( \sigma\chi \nabla x )  + \beta y\f]
-     * This is in principle possible also with the \c set_chi() and \c symv() functions
-     * however sometimes you have a \c sigma with explicit zeros or negative values.
-     * Then you need to use this function because \c set_chi() won't allow a \c sigma with zeros
-     * @note This function does not change the internal \c chi tensor
-     * @param alpha a scalar
-     * @param sigma The prefactor for the \c chi tensor
-     * @param x left-hand-side
-     * @param beta a scalar
-     * @param y result
-     * @tparam ContainerTypes must be usable with \c Container in \ref dispatch
-     */
-    template<class ContainerType0, class ContainerType1, class ContainerType2>
-    void multiply_sigma( value_type alpha, const ContainerType2& sigma, const ContainerType0& x, value_type beta, ContainerType1& y)
-    {
         //compute gradient
         dg::blas2::gemv( m_rightx, x, m_tempx); //R_x*f
         dg::blas2::gemv( m_righty, x, m_tempy); //R_y*f
 
         //multiply with tensor (note the alias)
-        dg::tensor::multiply2d(sigma, m_chi, m_tempx, m_tempy, 0., m_tempx, m_tempy);
+        dg::tensor::multiply2d(m_sigma, m_chi, m_tempx, m_tempy, 0., m_tempx, m_tempy);
 
         //now take divergence
         dg::blas2::symv( m_lefty, m_tempy, m_temp);
@@ -283,7 +284,7 @@ class Elliptic
             {
                 dg::blas2::symv( m_jfactor, m_jumpX, x, 0., m_tempx);
                 dg::blas2::symv( m_jfactor, m_jumpY, x, 0., m_tempy);
-                dg::tensor::multiply2d(sigma, m_chi, m_tempx, m_tempy, 0., m_tempx, m_tempy);
+                dg::tensor::multiply2d(m_sigma, m_chi, m_tempx, m_tempy, 0., m_tempx, m_tempy);
                 dg::blas1::axpbypgz(1.0,m_tempx,1.0,m_tempy,1.0,m_temp);
             }
             else
@@ -297,6 +298,49 @@ class Elliptic
         if( m_no == not_normed)//multiply weights without volume
             dg::blas1::pointwiseDot( alpha, m_weights_wo_vol, m_temp, beta, y);
     }
+
+    /**
+     * @brief \f$ \sigma = (\nabla\phi\cdot\chi\cdot\nabla \phi) \f$
+     *
+     * @param phi the vector to take the variation of
+     * @param sigma (inout) the variation
+     * @tparam ContainerTypes must be usable with \c Container in \ref dispatch
+     */
+    template<class ContainerType0, class ContainerType1>
+    void variation(const ContainerType0& phi, ContainerType1& sigma){
+        variation(1., 1., phi, 0., sigma);
+    }
+    /**
+     * @brief \f$ \sigma = \lambda^2(\nabla\phi\cdot\chi\cdot\nabla \phi) \f$
+     *
+     * @param lambda input prefactor
+     * @param phi the vector to take the variation of
+     * @param sigma (out) the variation
+     * @tparam ContainerTypes must be usable with \c Container in \ref dispatch
+     */
+    template<class ContainerTypeL, class ContainerType0, class ContainerType1>
+    void variation(const ContainerTypeL& lambda, const ContainerType0& phi, ContainerType1& sigma){
+        variation(1.,lambda, phi, 0., sigma);
+    }
+    /**
+     * @brief \f$ \sigma = \alpha \lambda^2 (\nabla\phi\cdot\chi\cdot\nabla \phi) + \beta \sigma\f$
+     *
+     * @param alpha scalar input prefactor
+     * @param lambda input prefactor
+     * @param phi the vector to take the variation of
+     * @param beta the output prefactor
+     * @param sigma (inout) the variation
+     * @tparam ContainerTypes must be usable with \c Container in \ref dispatch
+     */
+    template<class ContainerTypeL, class ContainerType0, class ContainerType1>
+    void variation(value_type alpha, const ContainerTypeL& lambda, const ContainerType0& phi, value_type beta, ContainerType1& sigma)
+    {
+        dg::blas2::gemv( m_rightx, phi, m_tempx); //R_x*f
+        dg::blas2::gemv( m_righty, phi, m_tempy); //R_y*f
+        dg::tensor::scalar_product2d(alpha, lambda, m_tempx, m_tempy, m_chi, lambda, m_tempx, m_tempy, beta, sigma);
+    }
+
+
     /**
      * @brief Determine if weights are multiplied to make operator symmetric or not
      *
@@ -323,20 +367,26 @@ using Elliptic2d = Elliptic<Geometry, Matrix, Container>;
 
 //Elliptic3d is tested in inc/geometries/elliptic3d_t.cu
 /**
- * @brief A 3d negative elliptic differential operator \f$ -\nabla \cdot ( \chi \nabla ) \f$
+ * @brief A 3d negative elliptic differential operator \f$ -\nabla \cdot ( \mathbf{\chi}\cdot \nabla ) \f$
  *
  * @ingroup matrixoperators
  *
- * The term discretized is \f[ -\nabla \cdot ( \mathbf \chi\cdot \nabla ) \f]
- * where \f$ \mathbf \chi \f$ is a positive semi-definit tensor.
- * In general coordinates that means
+ * The term discretized is \f[ -\nabla \cdot ( \mathbf{\chi} \cdot \nabla ) \f] where \f$
+ * \nabla \f$ is the two-dimensional nabla and \f$\chi = \sigma
+ * \mathbf{\tau}\f$ is a (possibly spatially dependent) tensor with scalar part
+ * \f$ \sigma\f$ (usually the volume form) and tensor part \f$ \tau\f$ (usually
+ * the inverse metric). In general coordinates that means
  * \f[ -\frac{1}{\sqrt{g}}\left(
  * \partial_x\left(\sqrt{g}\left(\chi^{xx}\partial_x + \chi^{xy}\partial_y + \chi^{xz}\partial_z \right)\right)
  + \partial_y\left(\sqrt{g}\left(\chi^{yx}\partial_x + \chi^{yy}\partial_y + \chi^{yz}\partial_z \right)\right)
  + \partial_z\left(\sqrt{g}\left(\chi^{zx}\partial_x + \chi^{zy}\partial_y + \chi^{zz}\partial_z \right)\right)
  \right)\f]
- is discretized. Note that the local discontinuous Galerkin discretization adds so-called
- jump terms
+ is discretized.
+ Per default, \f$ \chi = \sqrt{g} g^{-1}\f$ but you can set it to any tensor
+ you like (in order for the operator to be invertible \f$\chi\f$ should be
+ symmetric and positive definite though).
+
+ Note that the local discontinuous Galerkin discretization adds so-called jump terms
  \f[ D^\dagger \chi D + \alpha\chi_{on/off} J \f]
  where \f$\alpha\f$  is a scale factor ( = jfactor), \f$ D \f$ contains the discretizations of the above derivatives, and \f$ J\f$ is a self-adjoint matrix.
  (The symmetric part of \f$J\f$ is added @b before the volume element is divided). The adjoint of a matrix is defined with respect to the volume element including dG weights.
@@ -348,16 +398,18 @@ using Elliptic2d = Elliptic<Geometry, Matrix, Container>;
  The \f$ \chi_{on/off} \f$ in the jump term serves to weight the jump term with \f$ \chi \f$. This can be switched either on or off with off being the default.
  Also note that a forward discretization has more diffusion than a centered discretization.
 
- The following code snippet demonstrates the use of \c Elliptic in an inversion problem
+ The following code snippet demonstrates the use of \c Elliptic3d in an inversion problem
  * @snippet elliptic_b.cu invert
  * @copydoc hide_geometry_matrix_container
  * This class has the \c SelfMadeMatrixTag so it can be used in \c blas2::symv functions
  * and thus in a conjugate gradient solver.
- * @note The constructors initialize \f$ \chi=g\f$ so that a negative laplacian operator
- * results
+ * @note The constructors initialize \f$ \chi=\sqrt{g}g^{-1}\f$ so that a
+ * negative laplacian operator results
+ * @note The inverse of \f$ \sigma\f$ makes a good general purpose preconditioner
  * @note the jump term \f$ \alpha J\f$  adds artificial numerical diffusion as discussed above
+ * @note Since the pattern arises quite often (because of the ExB velocity \f$ u_E^2\f$ in the ion gyro-centre potential)
+ * this class also can compute the variation integrand \f$ \lambda^2\nabla \phi\cdot \chi\cdot\nabla\phi\f$
  * @attention Pay attention to the negative sign which is necessary to make the matrix @b positive @b definite
- *
  */
 template <class Geometry, class Matrix, class Container>
 class Elliptic3d
@@ -420,10 +472,8 @@ class Elliptic3d
         dg::assign( dg::create::inv_weights(g),   m_precond);
         m_temp = m_tempx = m_tempy = m_tempz = m_inv_weights;
         m_chi=g.metric();
-        m_vol=dg::tensor::volume(m_chi);
-        dg::tensor::scal( m_chi, m_vol);
+        m_sigma = m_vol = dg::tensor::volume(m_chi);
         dg::assign( dg::create::weights(g), m_weights_wo_vol);
-        dg::assign( dg::evaluate(dg::one, g), m_sigma);
     }
     ///@copydoc Elliptic::construct()
     template<class ...Params>
@@ -433,43 +483,23 @@ class Elliptic3d
         *this = Elliptic3d( std::forward<Params>( ps)...);
     }
 
-    /**
-     * @brief Change scalar part in Chi tensor
-     *
-     * Internally, we split the tensor \f$\chi = \sigma\mathbf{\tau}\f$ into
-     * a scalar part \f$ \sigma\f$ and a tensor part \f$ \tau\f$ and you can
-     * set each part seperately. This functions sets the scalar part.
-     *
-     * @param sigma The new scalar part in \f$\chi\f$ (all elements must be >0)
-     * @tparam ContainerType0 must be usable with \c Container in \ref dispatch
-     */
+    ///@copydoc Elliptic2d::set_chi(const ContainerType0&)
     template<class ContainerType0>
     void set_chi( const ContainerType0& sigma)
     {
-        dg::blas1::pointwiseDivide( sigma, m_sigma, m_tempx);
+        dg::blas1::pointwiseDot( sigma, m_vol, m_sigma);
         //update preconditioner
-        dg::blas1::pointwiseDivide( m_precond, m_tempx, m_precond);
-        dg::tensor::scal( m_chi, m_tempx);
-        dg::blas1::copy( sigma, m_sigma);
+        dg::blas1::pointwiseDivide( m_inv_weights, sigma, m_precond);
+        // sigma is possibly zero, which will invalidate the preconditioner
+        // it is important to call this blas1 function because it can
+        // overwrite NaN in m_precond in the next update
     }
-    /**
-     * @brief Change tensor part in Chi tensor
-     *
-     * Internally, we split the tensor \f$\chi = \sigma\mathbf{\tau}\f$ into
-     * a scalar part \f$ \sigma\f$ and a tensor part \f$ \tau\f$ and you can
-     * set each part seperately. This functions sets the tensor part.
-     *
-     * @note The class will take care of the volume element in the divergence so do not multiply it to \c tau yourself
-     *
-     * @param tau The new tensor part in \f$\chi\f$ (must be positive definite)
-     * @tparam ContainerType0 must be usable in \c dg::assign to \c Container
-     */
+
+    ///@copydoc Elliptic2d::set_chi(const SparseTensor<ContainerType0>&)
     template<class ContainerType0>
     void set_chi( const SparseTensor<ContainerType0>& tau)
     {
         m_chi = SparseTensor<Container>(tau);
-        dg::tensor::scal( m_chi, m_sigma);
-        dg::tensor::scal( m_chi, m_vol);
     }
 
     ///@copydoc Elliptic::inv_weights()
@@ -513,12 +543,6 @@ class Elliptic3d
     template<class ContainerType0, class ContainerType1>
     void symv( value_type alpha, const ContainerType0& x, value_type beta, ContainerType1& y)
     {
-        multiply_sigma( alpha, 1., x, beta, y);
-    }
-    ///@copydoc Elliptic::multiply_sigma(value_type,const ContainerType2&,const ContainerType0&,value_type,ContainerType1&)
-    template<class ContainerType0, class ContainerType1, class ContainerType2>
-    void multiply_sigma( value_type alpha, const ContainerType2& sigma, const ContainerType0& x, value_type beta, ContainerType1& y)
-    {
         //compute gradient
         dg::blas2::gemv( m_rightx, x, m_tempx); //R_x*f
         dg::blas2::gemv( m_righty, x, m_tempy); //R_y*f
@@ -527,14 +551,14 @@ class Elliptic3d
             dg::blas2::gemv( m_rightz, x, m_tempz); //R_z*f
 
             //multiply with tensor (note the alias)
-            dg::tensor::multiply3d(sigma, m_chi, m_tempx, m_tempy, m_tempz, 0., m_tempx, m_tempy, m_tempz);
+            dg::tensor::multiply3d(m_sigma, m_chi, m_tempx, m_tempy, m_tempz, 0., m_tempx, m_tempy, m_tempz);
             //now take divergence
             dg::blas2::symv( -1., m_leftz, m_tempz, 0., m_temp);
             dg::blas2::symv( -1., m_lefty, m_tempy, 1., m_temp);
         }
         else
         {
-            dg::tensor::multiply2d(sigma, m_chi, m_tempx, m_tempy, 0., m_tempx, m_tempy);
+            dg::tensor::multiply2d(m_sigma, m_chi, m_tempx, m_tempy, 0., m_tempx, m_tempy);
             dg::blas2::symv( -1.,m_lefty, m_tempy, 0., m_temp);
         }
         dg::blas2::symv( -1., m_leftx, m_tempx, 1., m_temp);
@@ -546,7 +570,7 @@ class Elliptic3d
             {
                 dg::blas2::symv( m_jfactor, m_jumpX, x, 0., m_tempx);
                 dg::blas2::symv( m_jfactor, m_jumpY, x, 0., m_tempy);
-                dg::tensor::multiply2d(m_chi, m_tempx, m_tempy, m_tempx, m_tempy);
+                dg::tensor::multiply2d(m_sigma, m_chi, m_tempx, m_tempy, 0., m_tempx, m_tempy);
                 dg::blas1::axpbypgz(1.0,m_tempx,1.0,m_tempy,1.0,m_temp);
             }
             else
@@ -561,11 +585,30 @@ class Elliptic3d
             dg::blas1::pointwiseDot( alpha, m_weights_wo_vol, m_temp, beta, y);
     }
 
-    /**
-     * @brief Determine if weights are multiplied to make operator symmetric or not
-     *
-     * @param new_norm new setting
-     */
+    ///@copydoc Elliptic::variation(const ContainerType0&,ContainerType1&)
+    template<class ContainerType0, class ContainerType1>
+    void variation(const ContainerType0& phi, ContainerType1& sigma){
+        variation(1.,1., phi, 0., sigma);
+    }
+    ///@copydoc Elliptic::variation(const ContainerTypeL&,const ContainerType0&,ContainerType1&){
+    template<class ContainerTypeL, class ContainerType0, class ContainerType1>
+    void variation(const ContainerTypeL& lambda, const ContainerType0& phi, ContainerType1& sigma){
+        variation(1.,lambda, phi, 0., sigma);
+    }
+    ///@copydoc Elliptic::variation(value_type,const ContainerTypeL&,const ContainerType0&,value_type,ContainerType1&)
+    template<class ContainerTypeL, class ContainerType0, class ContainerType1>
+    void variation(value_type alpha, const ContainerTypeL& lambda, const ContainerType0& phi, value_type beta, ContainerType1& sigma)
+    {
+        dg::blas2::gemv( m_rightx, phi, m_tempx); //R_x*f
+        dg::blas2::gemv( m_righty, phi, m_tempy); //R_y*f
+        if( m_multiplyZ)
+            dg::blas2::gemv( m_rightz, phi, m_tempz); //R_y*f
+        else
+            dg::blas1::scal( m_tempz, 0.);
+        dg::tensor::scalar_product3d(alpha, lambda,  m_tempx, m_tempy, m_tempz, m_chi, lambda, m_tempx, m_tempy, m_tempz, beta, sigma);
+    }
+
+    ///@copydoc Elliptic::set_norm(dg::norm)
     void set_norm( dg::norm new_norm) {
         m_no = new_norm;
     }
