+++ inc/dg/elliptic2d_b.cu
@@ -27,7 +27,9 @@ double rhs( double x, double y) { return 2.*sin(x)*sin(y)*(amp*sin(x)*sin(y)+1)-
 //double rhs( double x, double y) { return 2.*sin( x)*sin(y);}
 //double rhs( double x, double y) { return 2.*sin(x)*sin(y)*(sin(x)*sin(y)+1)-sin(x)*sin(x)*cos(y)*cos(y)-cos(x)*cos(x)*sin(y)*sin(y)+(x*sin(x)-cos(x))*sin(y) + x*sin(x)*sin(y);}
 double sol(double x, double y)  { return sin( x)*sin(y);}
-double der(double x, double y)  { return cos( x)*sin(y);}
+double derX(double x, double y)  { return cos( x)*sin(y);}
+double derY(double x, double y)  { return sin( x)*cos(y);}
+double vari(double x, double y)  { return pol(x,y)*pol(x,y)*(derX(x,y)*derX(x,y) + derY(x,y)*derY(x,y));}
 
 
 int main()
@@ -44,9 +46,9 @@ int main()
 	/*std::cout << "Type n, Nx and Ny and epsilon and jfactor (1)! \n";
     std::cin >> n >> Nx >> Ny; //more N means less iterations for same error
     std::cin >> eps >> jfactor;*/
-    bool jump_weight;
-    std::cout << "Jump weighting on or off? Type 1 for true or 0 for false: \n";
-    std::cin >> jump_weight;
+    bool jump_weight = false;
+    //std::cout << "Jump weighting on or off? Type 1 for true or 0 for false (default): \n";
+    //std::cin >> jump_weight;
     std::cout << "Computation on: "<< n <<" x "<< Nx <<" x "<< Ny << std::endl;
     //std::cout << "# of 2d cells                 "<< Nx*Ny <<std::endl;
 
@@ -63,7 +65,8 @@ int main()
     dg::DVec temp = x;
     //compute error
     const dg::DVec solution = dg::evaluate( sol, grid);
-    const dg::DVec derivati = dg::evaluate( der, grid);
+    const dg::DVec derivati = dg::evaluate( derX, grid);
+    const dg::DVec variatio = dg::evaluate( vari, grid);
     const double norm = dg::blas2::dot( w2d, solution);
     dg::DVec error( solution);
     dg::exblas::udouble res;
@@ -90,6 +93,11 @@ int main()
     for(unsigned u=0; u<stages; u++)
     {
         multi_pol[u].construct( multigrid.grid(u), dg::not_normed, dg::centered, jfactor);
+        //this tests if elliptic can recover from NaN in the preconditioner
+        multi_pol[u].set_chi(0.);
+        // here we test if we can set the tensor part in elliptic
+        multi_pol[u].set_chi( multigrid.grid(u).metric());
+        // now set the actual scalar part in chi
         multi_pol[u].set_chi( multi_chi[u]);
         multi_pol[u].set_jump_weighting(jump_weight);
     }
@@ -110,18 +118,18 @@ int main()
     double err = dg::blas2::dot( w2d, error);
     err = sqrt( err/norm); res.d = err;
     std::cout << " "<<err << "\t"<<res.i<<"\n";
-    }
-
     dg::DMatrix DX = dg::create::dx( grid);
     dg::blas2::gemv( DX, x, error);
     dg::blas1::axpby( 1.,derivati,-1., error);
-    double err = dg::blas2::dot( w2d, error);
+    err = dg::blas2::dot( w2d, error);
     const double norm_der = dg::blas2::dot( w2d, derivati);
-    std::cout << "L2 Norm of relative error in derivative is "<<std::setprecision(16)<< sqrt( err/norm_der)<<std::endl;
+    std::cout << "L2 Norm of relative error in derivative is\n "<<std::setprecision(16)<< sqrt( err/norm_der)<<std::endl;
     //derivative converges with p-1, for p = 1 with 1/2
 
+    }
+
     {
-        std::cout << "Forward Elliptic\n";
+    std::cout << "Forward Elliptic\n";
     x = temp;
     //![invert]
     //create an Elliptic object without volume form (not normed)
@@ -145,6 +153,12 @@ int main()
     //output the relative error
     std::cout << " "<<sqrt( err/norm) << "\n";
     //![invert]
+    std::cout << "Compute variation in forward Elliptic\n";
+    pol_forward.variation( 1., chi, x, 0., error);
+    dg::blas1::axpby( 1., variatio, -1., error);
+    err = dg::blas2::dot( w2d, error);
+    norm = dg::blas2::dot( w2d, variatio);
+    std::cout << " "<<sqrt( err/norm) << "\n";
     }
 
     {
