+++ inc/dg/gradient_t.cu
@@ -3,7 +3,6 @@
 
 #include "gradient.h"
 
-//![function]
 const double lx = 2*M_PI;
 const double ly = 2*M_PI;
 dg::bc bcx = dg::PER;
@@ -18,9 +17,6 @@ double dxphi( double x, double y) {
 double dyphi( double x, double y) {
     return -sin(x)*sin(y);
 }
-double variation( double x, double y) {
-    return dxphi(x,y)*dxphi(x,y) + dyphi(x,y)*dyphi(x,y);
-}
 double phi3d( double x, double y,double z) {
     return sin(x)*cos(y)*cos(z);
 }
@@ -33,16 +29,11 @@ double dyphi3d( double x, double y, double z) {
 double dzphi3d( double x, double y, double z) {
     return -sin(x)*cos(y)*sin(z)/x/x;
 }
-double variation3d( double x, double y, double z) {
-    return dxphi3d(x,y,z)*dxphi3d(x,y,z)
-        + dyphi3d(x,y,z)*dyphi3d(x,y,z)
-        + dzphi3d(x,y,z)*dzphi3d(x,y,z)*x*x;
-}
 
 // There are more tests in geometries/geometry_advection_(mpi)b.cu
 int main()
 {
-    std::cout<<"This program tests the execution of the gradient and variation scheme! A test is passed if the number in the second column shows exactly zero!\n";
+    std::cout<<"This program tests the execution of the gradient! A test is passed if the number in the second column shows exactly zero!\n";
     unsigned n = 5, Nx = 32, Ny = 48;
     std::cout << "TEST 2D\n";
     std::cout << "Computing on the Grid " <<n<<" x "<<Nx<<" x "<<Ny <<std::endl;
@@ -54,17 +45,15 @@ int main()
     const dg::DVec ph = dg::construct<dg::DVec>( dg::evaluate( phi, grid));
     const dg::DVec phx = dg::construct<dg::DVec>( dg::evaluate( dxphi, grid));
     const dg::DVec phy = dg::construct<dg::DVec>( dg::evaluate( dyphi, grid));
-    const dg::DVec var = dg::construct<dg::DVec>( dg::evaluate( variation, grid));
-    dg::DVec dxph(ph), dyph(ph), va(ph);
+    dg::DVec dxph(ph), dyph(ph);
 
     // create a Gradient object
     dg::Gradient<dg::aGeometry2d, dg::DMatrix, dg::DVec> gradient( grid);
 
     //apply arakawa scheme
     gradient.gradient( ph, dxph, dyph);
-    gradient.variation( ph, va);
 
-    int64_t binary[] = {4500635718861276907,4487444521638156650,4499885861996435701};
+    int64_t binary[] = {4500635718861276907,4487444521638156650};
     dg::exblas::udouble res;
     dg::DVec w2d = dg::create::weights( grid);
 
@@ -74,9 +63,6 @@ int main()
     dg::blas1::axpby( 1., phy, -1., dyph);
     res.d = sqrt(dg::blas2::dot( w2d, dyph)); //don't forget sqrt when computing errors
     std::cout << "Gy Distance to solution "<<res.d<<"\t\t"<<res.i-binary[1]<<std::endl;
-    dg::blas1::axpby( 1., var, -1., va);
-    res.d = sqrt(dg::blas2::dot( w2d, va)); //don't forget sqrt when computing errors
-    std::cout << "V  Distance to solution "<<res.d<<"\t\t"<<res.i-binary[2]<<std::endl;
     //periocid bc       |  dirichlet bc
     //n = 1 -> p = 2    |
     //n = 2 -> p = 1    |
@@ -95,17 +81,15 @@ int main()
     const dg::DVec phx3d = dg::construct<dg::DVec>( dg::evaluate( dxphi3d, grid3d));
     const dg::DVec phy3d = dg::construct<dg::DVec>( dg::evaluate( dyphi3d, grid3d));
     const dg::DVec phz3d = dg::construct<dg::DVec>( dg::evaluate( dzphi3d, grid3d));
-    const dg::DVec var3d = dg::construct<dg::DVec>( dg::evaluate( variation3d, grid3d));
-    dg::DVec dxph3d(ph3d), dyph3d(ph3d), dzph3d(ph3d), va3d(ph3d);
+    dg::DVec dxph3d(ph3d), dyph3d(ph3d), dzph3d(ph3d);
 
     // create a Gradient object
     dg::Gradient3d<dg::aGeometry3d, dg::DMatrix, dg::DVec> gradient3d( grid3d);
 
     //apply arakawa scheme
     gradient3d.gradient( ph3d, dxph3d, dyph3d, dzph3d);
-    gradient3d.variation( ph3d, va3d);
 
-    int64_t binary3d[] = {4504451755369532568,4491224193368827475,4549042274897523598,4550331496568322612};
+    int64_t binary3d[] = {4504451755369532568,4491224193368827475,4549042274897523598};
     dg::exblas::udouble res3d;
     dg::DVec w3d = dg::create::weights( grid3d);
 
@@ -118,9 +102,6 @@ int main()
     dg::blas1::axpby( 1., phz3d, -1., dzph3d);
     res3d.d = sqrt(dg::blas2::dot( w3d, dzph3d)); //don't forget sqrt when computing errors
     std::cout << "Gz Distance to solution "<<res3d.d<<"\t\t"<<res3d.i-binary3d[2]<<std::endl;
-    dg::blas1::axpby( 1., var3d, -1., va3d);
-    res3d.d = sqrt(dg::blas2::dot( w3d, va3d)); //don't forget sqrt when computing errors
-    std::cout << "V  Distance to solution "<<res3d.d<<"\t\t"<<res3d.i-binary3d[3]<<std::endl;
     //periocid bc       |  dirichlet bc
     //n = 1 -> p = 2    |
     //n = 2 -> p = 1    |
