+++ src/ep/toeflR.cuh
@@ -125,7 +125,6 @@ struct ToeflR
     dg::Elliptic<Geometry, Matrix, container> pol, laplaceM; //contains normalized laplacian
     dg::Helmholtz<Geometry,  Matrix, container> gamma1;
     dg::ArakawaX< Geometry, Matrix, container> arakawa; 
-    dg::Variation< Geometry, Matrix, container> gradient; 
 
     dg::Invert<container> invert_pol, invert_invgamma;
 
@@ -150,7 +149,6 @@ ToeflR< Geometry, M, container>::ToeflR( const Geometry& grid, const Parameters&
     laplaceM( grid, dg::normed, dg::centered),
     gamma1(  grid, 0., dg::centered),
     arakawa( grid),
-    gradient( grid, dg::centered),
     invert_pol(      omega, omega.size(), p.eps_pol),
     invert_invgamma( omega, omega.size(), p.eps_gamma),
     w2d( dg::create::volume(grid)), v2d( dg::create::inv_volume(grid)), one( dg::evaluate(dg::one, grid)),
@@ -171,7 +169,7 @@ void ToeflR<G, M, container>::compute_psi( const container& phi)
         if(  number == invert_invgamma.get_max())
             throw dg::Fail( eps_gamma);
 
-        gradient.variation(binv, phi, omega); //needed also in local energy theorem
+        pol.variation(binv, phi, omega); //needed also in local energy theorem
         dg::blas1::axpby( 1., psi[i], -0.5*mu[i], omega, psi[i]);   //psi  Gamma phi - 0.5 u_E^2
     }
 }
@@ -233,7 +231,7 @@ void ToeflR<G, M, container>::operator()(double t, const std::vector<container>&
         double Ue = z[0]*tau[0]*dg::blas2::dot( lny[0], w2d, ype[0]);
         double Up = z[1]*tau[1]*dg::blas2::dot( lny[1], w2d, ype[1]);
         double Uphi = 0.5*dg::blas2::dot( ype[0], w2d, omega) + 0.5*dg::blas2::dot( ype[1], w2d, omega); 
-        gradient.variation(potential_, omega); 
+        pol.variation(potential_, omega); 
         double UE = debye_*dg::blas2::dot( one, w2d, omega);
         energy_ = Ue + Up + Uphi + UE;
         //std::cout << "Ue "<<Ue<< "Up "<<Up<< "Uphi "<<Uphi<< "UE "<<UE<<"\n";
